Support the case where there are multiple instances of a given agent
A smith agency can have multiple instance of the same agent, and in states
other than "running".  The previous version of the monitoring agent would
only "see" the instance of a given agent that was last in the returned list.
If that instance happened to be non-running (starting or stopping), then the
agent would consider that agent to be failed, even though there actually
*was* an instance of that agent happily chugging away.

Now, we parse out all instances of the agent from the returned list, and the
liveness check looks for at least one instance that is running and has its
process in the process table.